+++ src/main/java/com/example/oatnote/domain/memotag/AsyncMemoTagService.java
@@ -1,6 +1,5 @@
 package com.example.oatnote.domain.memotag;
 
-import java.time.LocalDateTime;
 import java.util.*;
 
 import org.redisson.api.RLock;
@@ -8,10 +7,10 @@ import org.redisson.api.RedissonClient;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
-import com.example.oatnote.domain.memotag.service.client.AiMemoTagClient;
-import com.example.oatnote.domain.memotag.service.client.dto.AiCreateStructureRequest;
-import com.example.oatnote.domain.memotag.service.client.dto.AiCreateStructureResponse;
-import com.example.oatnote.domain.memotag.service.client.dto.AiCreateTagsResponse;
+import com.example.oatnote.domain.memotag.service.aiClient.AiMemoTagClient;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.AiCreateStructureRequest;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.AiCreateStructureResponse;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.AiCreateTagsResponse;
 import com.example.oatnote.domain.memotag.service.memo.MemoService;
 import com.example.oatnote.domain.memotag.service.memo.model.Memo;
 import com.example.oatnote.domain.memotag.service.relation.MemoTagRelationService;
