+++ src/main/java/com/example/oatnote/domain/memotag/dto/CreateMemoRequest.java
@@ -5,10 +5,9 @@ import static com.example.oatnote.web.validation.enums.AllowedFileTypeEnum.IMAGE
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
-import com.example.oatnote.domain.memotag.service.client.dto.AiCreateTagsRequest;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.AiCreateTagsRequest;
 import com.example.oatnote.domain.memotag.service.memo.model.Memo;
 import com.example.oatnote.web.validation.AllowedFileType;
 import com.example.oatnote.web.validation.MemoAtLeastOneRequired;
