+++ src/main/java/com/example/oatnote/domain/memotag/dto/UpdateMemoTagsResponse.java
@@ -3,7 +3,7 @@ package com.example.oatnote.domain.memotag.dto;
 import java.util.List;
 
 import com.example.oatnote.domain.memotag.dto.innerDto.MemoResponse;
-import com.example.oatnote.domain.memotag.service.client.dto.innerDto.RawTag;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.innerDto.RawTag;
 import com.example.oatnote.domain.memotag.service.memo.model.Memo;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
