+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/AiCreateEmbeddingResponse.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.domain.memotag.service.client.dto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto;
 
 import java.util.List;
 
@@ -6,9 +6,8 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record AiCreateMetadataResponse(
-    String metadata,
-    List<Double> embeddingMetadata
+public record AiCreateEmbeddingResponse(
+    List<Double> embedding
 ) {
 
 }
