+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/AiCreateMetadataResponse.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.domain.memotag.service.client.dto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto;
 
 import java.util.List;
 
@@ -6,8 +6,9 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record AiCreateEmbeddingResponse(
-    List<Double> embedding
+public record AiCreateMetadataResponse(
+    String metadata,
+    List<Double> embeddingMetadata
 ) {
 
 }
