+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/AiCreateTagsResponse.java
@@ -1,9 +1,9 @@
-package com.example.oatnote.domain.memotag.service.client.dto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.example.oatnote.domain.memotag.service.client.dto.innerDto.RawTag;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.innerDto.RawTag;
 import com.example.oatnote.domain.memotag.service.memo.model.Memo;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
