+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/AiSearchMemosUsingAiRequest.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.domain.memotag.service.client.dto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto;
 
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
