+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/AiSearchMemosUsingAiResponse.java
@@ -1,10 +1,10 @@
-package com.example.oatnote.domain.memotag.service.client.dto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto;
 
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 
 import java.util.List;
 
-import com.example.oatnote.domain.memotag.service.client.dto.enums.AiSearchTypeEnum;
+import com.example.oatnote.domain.memotag.service.aiClient.dto.enums.AiSearchTypeEnum;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(SnakeCaseStrategy.class)
