+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/AiSearchMemosUsingDbRequest.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.domain.memotag.service.client.dto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
