+++ src/main/java/com/example/oatnote/domain/memotag/service/aiClient/dto/innerDto/ProcessedMemoTags.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.domain.memotag.service.client.dto.innerDto;
+package com.example.oatnote.domain.memotag.service.aiClient.dto.innerDto;
 
 import java.time.LocalDateTime;
 import java.util.List;
